Move all the formula handling to the formula sequence
