+++ src/clients/configuration_management_formula_ng.rb
@@ -1,3 +1,3 @@
 require "y2configuration_management/clients/formula"
 
-Y2ConfigurationManagement::Clients::Formula.new.main
+Y2ConfigurationManagement::Clients::Formula.new.run
