+++ src/lib/y2configuration_management/clients/formula.rb
@@ -18,6 +18,8 @@
 # find current contact information at www.suse.com.
 
 require "yast"
+require "configuration_management/configurators/salt"
+require "configuration_management/configurations/salt"
 require "y2configuration_management/salt/formula"
 require "y2configuration_management/salt/formula_sequence"
 require "configuration_management/cfa/salt_top"
@@ -25,61 +27,32 @@ require "configuration_management/cfa/salt_top"
 module Y2ConfigurationManagement
   module Clients
     # Client to configure formulas
-    class Formula < Yast::Client
-      include Yast::Logger
-      extend Yast::I18n
+    class Formula
 
-      # @return [Array<Y2ConfigurationManagement::Salt::Formula>]
-      attr_accessor :formulas
-      # @return [String]
-      attr_reader :states_root, :formulas_root, :pillar_root
-
-      # Constructor
-      def initialize
-        textdomain "configuration_management"
-      end
-
-      def main
-        configure_directories
-        read_formulas
-        start_workflow
-        write_formulas
+      def run
+        configurator = Yast::ConfigurationManagement::Configurators::Base.for(config)
+        configurator.prepare
+        Y2ConfigurationManagement::Salt::FormulaSequence.new(config).run
       end
 
     private
 
-      def start_workflow
-        Y2ConfigurationManagement::Salt::FormulaSequence.new(formulas).run
+      # Returns the configuration management configuration
+      #
+      # @return [Yast::ConfigurationManagement::Configurations::Base]
+      def config
+        return @config if @config
+        settings =
+          {
+            "type"           => "salt",
+            "mode"           => "masterless",
+            "formulas_roots" => Y2ConfigurationManagement::Salt::Formula.formula_directories,
+            "states_roots"   => Y2ConfigurationManagement::Salt::Formula::BASE_DIR + "/states",
+            "pillar_root"    => Y2ConfigurationManagement::Salt::Formula::DATA_DIR + "/pillar"
+          }
+        @config = Yast::ConfigurationManagement::Configurations::Base.import(settings)
       end
 
-      def configure_directories
-        @states_root, @formulas_root, @pillar_root = Yast::WFM.Args()
-        @states_root ||= Y2ConfigurationManagement::Salt::Formula::BASE_DIR + "/states"
-        @formulas_root ||= Y2ConfigurationManagement::Salt::Formula.formula_directories
-        @pillar_root ||= Y2ConfigurationManagement::Salt::Formula::DATA_DIR + "/pillar"
-      end
-
-      def read_formulas
-        self.formulas = Y2ConfigurationManagement::Salt::Formula.all(formulas_root)
-        formulas.each { |f| f.pillar = pillar_for(f) }
-      end
-
-      def write_formulas
-        [pillar_root, states_root].each do |path|
-          ::FileUtils.mkdir_p(path) unless File.exist?(path)
-          top = Yast::ConfigurationManagement::CFA::SaltTop.new(path: File.join(path, "top.sls"))
-          top.load
-          top.add_states(formulas.select(&:enabled?).map(&:id))
-          top.save
-        end
-      end
-
-      def pillar_for(formula)
-        pillar_file = File.join(pillar_root, "#{formula.id}.sls")
-        pillar = Y2ConfigurationManagement::Salt::Pillar.new(data: {}, path: pillar_file)
-        pillar.load
-        pillar
-      end
     end
   end
 end
